Alpha version of the Robotics Library with modified Api Seems to be stable
The ConnectionManager is no longer a fundamental piece for send data to the CommandManager,
now IConnector instances are used. The ConnectionManager property of the CommandManager is
kept for compatibility reasons, but within the ConnectionManager instance is wrapped by a
TcpPacketParser which extends IConnector.

Also a Sample project has been added to quick check functionality